+++ packages/go_router/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 5.0.0
+
+- Fixes a bug where intermediate route redirect methods are not called.
+- GoRouter implements the RouterConfig interface, allowing you to call
+  MaterialApp.router(routerConfig: _myGoRouter) instead of passing
+  the RouterDelegate, RouteInformationParser, and RouteInformationProvider
+  fields.
+- **BREAKING CHANGE**
+  - Redesigns redirection API, adds asynchronous feature, and adds build context to redirect.
+  - Removes GoRouterRefreshStream
+  - Removes navigatorBuilder
+  - Removes urlPathStrategy
+- [go_router v5 migration guide](https://flutter.dev/go/go-router-v5-breaking-changes)
+
 ## 4.5.1
 
 - Fixes an issue where GoRoutes with only a redirect were disallowed